\name{RcmdrPlugin.TeachingDemos-package}
\alias{RcmdrPlugin.TeachingDemos-package}
\alias{RcmdrPlugin.TeachingDemos}
\docType{package}
\title{
Install the Demos Rcmdr Plug-In
}
\description{
This package provides an Rcmdr "plug-in" based on the TeachingDemos package, and is primarily for illustrative purposes.
}
\details{
\tabular{ll}{
Package: \tab RcmdrPlugin.TeachingDemos\cr
Type: \tab Package\cr
Version: \tab 1.0-5\cr
Date: \tab 2010-05-18\cr
License: \tab GPL (>= 2)\cr
}
}

\author{
John Fox <jfox@mcmaster.ca>

Maintainer: John Fox <jfox@mcmaster.ca>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}

