\name{RcmdrDemosPlugin-package}
\alias{RcmdrDemosPlugin-package}
\alias{RcmdrDemosPlugin}
\docType{package}
\title{
Install the Demos Rcmdr Plug-In
}
\description{
This package provides an Rcmdr "plug-in" based on the TeachingDemos package, and is primarily for illustrative purposes.
}
\details{
\tabular{ll}{
Package: \tab RcmdrDemosPlugin\cr
Type: \tab Package\cr
Version: \tab 1.0-0\cr
Date: \tab 2007-05-18\cr
License: \tab GPL version 2 or newer\cr
}
}

\author{
John Fox <jfox@mcmaster.ca>

Maintainer: John Fox <jfox@mcmaster.ca>
}

\keyword{ package }
\seealso{
\code{\link[Rcmdr]{Rcmdr}}.
}

}
